pyproject.toml
[tool.poetry]
name = "summoner-spell-tracker"
version = "0.1.0"
description = ""
authors = ["Adhithya Bhaskar <adhithyabhaskar@gmail.com>"]
packages =  [{ include = "src" }]

[tool.poetry.dependencies]
python = "=3.9.16"
jupyterlab = "^3.6.3"
numpy = "^1.24.2"
pandas = "^2.0.0"
black = "^23.3.0"
mypy = "^1.2.0"
seaborn = "^0.12.2"
torch = "^2.0.0"
torchvision = "^0.15.1"
tqdm = "^4.65.0"
tensorboard = "^2.12.2"
h5py = "^3.8.0"
pytorch-model-summary = "^0.1.2"
torchsummary = "^1.5.1"
flashtorch = "^0.1.3"
torchviz = "^0.0.2"
pytorch-lightning = "^2.0.2"
trdg = "^1.8.0"
pytest = "^7.3.1"
wandb = "^0.15.0"
rich = "^13.3.5"
random-username = "^1.0.2"
ffprobe-python = "^1.0.3"
ffmpeg-python = "^0.2.0"
pytesseract = "^0.3.10"
opencv-python = "^4.7.0.72"
torchmetrics = "^0.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
